requirements.txt
# SESORA 依赖
# Python 3.9+

# 数据类反序列化
dacite>=1.8.0

# 阿里云 SDK
aliyun-python-sdk-core>=2.13.0
alibabacloud-rds20140815>=15.0.0
oss2>=2.18.0
aliyun-log-python-sdk>=0.7.0  # SLS 日志服务 SDK
aliyun-python-sdk-cms>=7.0.0
alibabacloud-cms20190101>=2.0.0
aliyun-python-sdk-arms>=2.0.0
alibabacloud-arms20190808>=10.0.0
alibabacloud-cr20181201>=1.0.0
alibabacloud_fc20230330>=1.0.0
alibabacloud-ros20190910>=1.0.0
alibabacloud-alb20200616>=1.0.0
alibabacloud-cs20151215>=6.6.0
alibabacloud-ecs20140526>=7.6.0
alibabacloud-devops20210625>=5.0.3
alibabacloud-sls20201230>=5.12.1

# Kubernetes 客户端
kubernetes>=28.0.0

# HTTP 客户端
aiohttp>=3.8.0
requests>=2.31.0

# YAML 解析
PyYAML>=6.0.1

# 环境变量管理
python-dotenv>=1.0.0
